% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareMETALS.range.multiAllele.R
\name{assocGene}
\alias{assocGene}
\title{Association analysis of gene-level test with multi-allele sites;}
\usage{
assocGene(raw.data, maf.cutoff, rvTest)
}
\arguments{
\item{raw.data}{raw.data from seqminer output;}

\item{maf.cutoff}{The minor allele freq cutoff}

\item{rvTest}{The rare variant association test}
}
\value{
A list of results;
}
\description{
Association analysis of gene-level test with multi-allele sites;
}

