% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareMETALS2.utils.R
\name{binom.conf}
\alias{binom.conf}
\title{get binomial confidence intervals via normal approx}
\usage{
binom.conf(n, p, alpha = 0.95)
}
\arguments{
\item{n}{sample size}

\item{p}{estimated proportions}

\item{alpha}{alpha level}
}
\value{
conf.upper and conf.lower
}
\description{
get binomial confidence intervals via normal approx
}
