% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvmeta.CMH.R
\name{calcGeneLevelTest}
\alias{calcGeneLevelTest}
\title{rvmeta.CMH.new which takes ustat, V stat as input;}
\usage{
calcGeneLevelTest(ustat, V, maf.vec, N, test = "GRANVIL")
}
\description{
ustat ustat
V the matrix of V;
maf.vec the vector of minor allele frequencies
N sample size;
test The rvtests. Choices are GRANVIL, SKAT or VT;
}
