% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{checkOverlap}
\alias{checkOverlap}
\title{check sample overlaps between cohorts;}
\usage{
checkOverlap(score.stat.file1, score.stat.file2, ...)
}
\value{
a list with estimated lambda values;
}
\description{
check sample overlaps between cohorts;
}
