% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareMETALS2.utils.R
\name{colCenter}
\alias{colCenter}
\title{This function implements function to colCenter a matrix so that colSums(matIn.centered) is 0;}
\usage{
colCenter(matIn)
}
\arguments{
\item{matIn}{Matrix input;}
}
\value{
colCentered matrix;
}
\description{
This function implements function to colCenter a matrix so that colSums(matIn.centered) is 0;
}
