% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional.rareMETALS.range.R
\name{conditional.rareMETALS.range}
\alias{conditional.rareMETALS.range}
\title{Perform conditional analysis for gene-level tests}
\usage{
conditional.rareMETALS.range(range.name = NULL, score.stat.file, cov.file,
  candidate.variant.vec, known.variant.vec, test = "GRANVIL", maf.cutoff,
  alternative = c("two.sided", "greater", "less"), ix.gold = 1,
  out.digits = 4, callrate.cutoff = 0, hwe.cutoff = 0, max.VT = NULL)
}
\arguments{
\item{range.name}{name of the range to be analyzed (for example, it can be a gene name e.g. APOE)}

\item{score.stat.file}{files of score statistics}

\item{cov.file}{covariance matrix files}

\item{candidate.variant.vec}{Vectors of candidate variants: e.g. c("1:123","1:1234"). The quotation is necessary!!}

\item{known.variant.vec}{Vectors of known variants: e.g. c("1:123","1:1234"). The quotation is necessary!!}

\item{test}{test of rare variant tests}

\item{maf.cutoff}{Cutoffs of MAF used for determining rare variants}

\item{alternative}{Alternative hypothesis to be tested}

\item{ix.gold}{Index of the gold standard population: Used for flipping alleles}

\item{out.digits}{The number of digits used in the output}

\item{callrate.cutoff}{Cutoff of call rates. Sites with callrates lower than the cutoff will be labeled as missing}

\item{hwe.cutoff}{Cutoff of HWE p-values. Sites with HWE pvalues lower than the cutoff will be labeled as missing}

\item{max.VT}{The maximum number of thresholds used in VT; Setting max.VT to 10 can improve the speed for calculation without affecting the power too much. The default parameter is NULL, which does not set upper limit on the number of variable frequency threhsold.}
}
\description{
Perform conditional analysis for gene-level tests
}

