% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional.rareMETALS.single.group.R
\name{conditional.rareMETALS.single.group}
\alias{conditional.rareMETALS.single.group}
\title{Perform conditional analysis for single variant tests}
\usage{
conditional.rareMETALS.single.group(candidate.variant, score.stat.file,
  cov.file, known.variant.vec, refaltList, maf.cutoff,
  alternative = c("two.sided", "greater", "less"), out.digits = 4,
  callrate.cutoff = 0, hwe.cutoff = 0, correctFlip = TRUE,
  knownCoding = "identity", impMissing = TRUE)
}
\arguments{
\item{candidate.variant}{Candidate variant position;}

\item{score.stat.file}{files of score statistics}

\item{cov.file}{covariance matrix files}

\item{known.variant.vec}{range of candidate variant, expressed in a vector, e.g. c("1:12345","1:234567");}

\item{refaltList}{A list of ref, alt alleles, as well as variant frequencies, whether it needs to check AF as for flipping alleles;}

\item{maf.cutoff}{Cutoffs of MAF used for determining rare variants}

\item{alternative}{Alternative hypothesis to be tested}

\item{out.digits}{The number of digits used in the output}

\item{callrate.cutoff}{Cutoff of call rates. Sites with callrates lower than the cutoff will be labeled as missing}

\item{hwe.cutoff}{Cutoff of HWE p-values. Sites with HWE pvalues lower than the cutoff will be labeled as missing}

\item{knownCoding}{knownCoding option allows users to specify if a linear transformation needs to be applied known variants. Possible choices for knownCoding option include identical, burden, randomm-eff.}

\item{test}{test of rare variant tests}
}
\value{
return a list of meta-analysis results
}
\description{
Perform conditional analysis for single variant tests
}

