% Generated by roxygen2 (4.0.0): do not edit by hand
\name{conditional.rareMETALS.single}
\alias{conditional.rareMETALS.single}
\title{Perform conditional analysis for single variant tests}
\usage{
conditional.rareMETALS.single(candidate.variant.vec, score.stat.file, cov.file,
  known.variant.vec, maf.cutoff, no.boot = 10000,
  alternative = c("two.sided", "greater", "less"), ix.gold = 1,
  out.digits = 4, callrate.cutoff = 0, hwe.cutoff = 0,
  p.value.known.variant.vec = NA, anno.known.variant.vec = NA,
  anno.candidate.variant.vec = NA)
}
\arguments{
\item{candidate.variant}{name of the range to be analyzed (for example, it can be a gene name e.g. APOE)}

\item{score.stat.file}{files of score statistics}

\item{cov.file}{covariance matrix files}

\item{known.variant.vec}{range of candidate variant, expressed in a vector, e.g. c("1:12345","1:234567");}

\item{known.variant.range}{range of known variant, expressed in tabix format, e.g. c("1:123456","1:2345678");}

\item{test}{test of rare variant tests}

\item{maf.cutoff}{Cutoffs of MAF used for determining rare variants}

\item{alternative}{Alternative hypothesis to be tested}

\item{out.digits}{The number of digits used in the output}

\item{callrate.cutoff}{Cutoff of call rates. Sites with callrates lower than the cutoff will be labeled as missing}

\item{hwe.cutoff}{Cutoff of HWE p-values. Sites with HWE pvalues lower than the cutoff will be labeled as missing}
}
\value{
return a list of meta-analysis results
}
\description{
Perform conditional analysis for single variant tests
}

