% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.gene.R
\name{find.top.variant}
\alias{find.top.variant}
\title{Find top variants in a gene region that satisfy a MAF cutoff;}
\usage{
find.top.variant(candidate.variant.vec, window.size, singlevar.result,
  pval.cutoff = 3.1e-07)
}
\arguments{
\item{window.size}{The size of window, default is 1e6}

\item{singlevar.result}{Single varinat association results;}

\item{pval.cutoff}{The cutoffs of p-values;}

\item{candidate.vairant.vec}{The vector of candidate variant;}
}
\description{
Find top variants in a gene region that satisfy a MAF cutoff;
}

