% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond.rvmeta.core.R
\name{get.conditional.score.stat}
\alias{get.conditional.score.stat}
\title{This is the function to obtain conditional score statistics;}
\usage{
get.conditional.score.stat(ustat, X.T.times.X, N, ix.candidate, ix.known)
}
\arguments{
\item{ustat}{The non-standardized score statistics}

\item{X.T.times.X}{The covariance matrix for ustat;}

\item{ix.candidate}{the indics of candidate variant;}

\item{ix.known}{the indices of known variants that are conditioned on;}
}
\value{
the score statistic and their covariance matrix;
}
\description{
This is the function to obtain conditional score statistics;
}

