% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.gene.R
\name{get.gene.inWindow}
\alias{get.gene.inWindow}
\title{Find nearby genes for a sigificant SNP}
\usage{
get.gene.inWindow(known.variant.vec, window.size = 1e+06)
}
\arguments{
\item{known.variant.vec}{Positions of known variants, in the format of 1:12345}

\item{window.size}{The window size. Genes within the window are extracted. Default size is 1e6}
}
\value{
A vector of genes;
}
\description{
Find nearby genes for a sigificant SNP
}

