% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{imputeConditional}
\alias{imputeConditional}
\title{Impute missing summary association statistics assuming}
\usage{
imputeConditional(ustat.list, vstat.list, cov.mat.list, N.mat,
  beta.vec = NULL, ix.candidate, ix.known)
}
\arguments{
\item{ustat.list}{the score statistics;}

\item{vstat.list}{the vstat list;}

\item{cov.mat.list}{the list of the covariance matrix}

\item{N.mat}{the matrix of sample sizes;each row for a study and each column for a variant site;}
}
\description{
Impute missing summary association statistics assuming
}

