% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{imputeMeta}
\alias{imputeMeta}
\title{Impute missing summary association statistics assuming}
\usage{
imputeMeta(ustat.list, vstat.list, cov.mat.list, N.mat, beta.vec = NULL)
}
\arguments{
\item{ustat.list}{the score statistics;}

\item{vstat.list}{the vstat list;}

\item{cov.mat.list}{the list of the covariance matrix}

\item{N.mat}{the matrix of sample sizes;each row for a study and each column for a variant site;}
}
\description{
Impute missing summary association statistics assuming
}

