% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareMETALS2.utils.R
\name{myFormat}
\alias{myFormat}
\title{format according to publication standard;}
\usage{
myFormat(n, digits)
}
\arguments{
\item{n}{The number to be formatted;}

\item{digits}{how many digits to be retained;}
}
\value{
formatted result a vector;
}
\description{
format according to publication standard;
}
