% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareMETALS2.utils.R
\name{paste.pos}
\alias{paste.pos}
\title{This function quickly paste chrom and pos;}
\usage{
paste.pos(chrom, pos)
}
\arguments{
\item{chrom}{Chromosome number;}

\item{pos}{Position}
}
\value{
chrom:pos;
}
\description{
This function quickly paste chrom and pos;
}

