% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myUtils.R
\name{pwrCalc}
\alias{pwrCalc}
\title{power calculation}
\usage{
pwrCalc(beta, maf, alpha, N = 3e+05)
}
\arguments{
\item{beta}{genetic effect sizes;}

\item{maf}{minor allele frequencies;}

\item{alpha}{significance threshold}

\item{N}{sample size}
}
\value{
power
}
\description{
power calculation
}
