% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareMETALS.calcGC.R
\name{rareMETALS.calcGC}
\alias{rareMETALS.calcGC}
\title{calculate GC by frequency bins;}
\usage{
rareMETALS.calcGC(score.stat.file, maf.bin)
}
\arguments{
\item{score.stat.file}{The file names for score statistics;}

\item{maf.bin}{The frequency bins. The GC for variants in each frequency bin will be calculated separately;}
}
\value{
A list which consist of GCs for each study;
}
\description{
calculate GC by frequency bins;
}
