% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rareMETALS.cleanScore}
\alias{rareMETALS.cleanScore}
\title{Single variant meta-analysis}
\usage{
rareMETALS.cleanScore(score.stat.file, refaltList, hwe.cutoff, callrate.cutoff,
  af.diff.max = 0.7, checkAF = TRUE)
}
\arguments{
\item{score.stat.file}{files of score statistics}

\item{callrate.cutoff}{Cutoffs of call rate, lower than which will NOT be analyzed (labelled as missing)}

\item{hwe.cutoff}{Cutoffs of HWE p-values}

\item{refaltList}{A list that contains pos, ref and alt for calibrating variant sites to have the right ref and alt alleles;}
}
\value{
a list consisting of results
}
\description{
Single variant meta-analysis
}

