% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareMETALS.estimatePVE.R
\name{rareMETALS.estimatePVE}
\alias{rareMETALS.estimatePVE}
\title{Estimate the variance explained for a list of variants (possibly in LD);}
\usage{
rareMETALS.estimatePVE(score.stat.file, cov.file, range, range.name, refaltList,
  maf.cutoff, out.digits, callrate.cutoff, hwe.cutoff, correctFlip = TRUE,
  lambda = 0)
}
\arguments{
\item{score.stat.file}{files of score statistics}

\item{cov.file}{Covariance matrix files}

\item{range}{tabix range for each gene/region; Tabix range needs to be in the format of "1:123-1234". Quotation marks are necessary.}

\item{range.name}{The name of the range,e.g. gene names can be used;}

\item{refaltList}{A list of reference, alternative allele, a vector of alternative allele frequencies; Specifically, the list should consist of}

\item{maf.cutoff}{MAF cutoff used to analyze variants; Default value is 1, i.e. no cutoffs are applied. MAFs are based upon the sample MAFs.}

\item{out.digits}{Number of digits used in the output}

\item{callrate.cutoff}{Cutoffs of call rate, lower than which will NOT be analyzed (labelled as missing)}

\item{hwe.cutoff}{Cutoffs of HWE p-values}

\item{correctFlip}{Correct for flipped sites for score statistics and their covariance matrices}

\item{lambda}{Tuning parameter to the variant the variance correlation matrix to avoid near singular LD matrix; Assume the orignal LD matrix is R2, R2.adj=R2+lambda*Id;}
}
\value{
a list consisting of results;
}
\description{
Estimate the variance explained for a list of variants (possibly in LD);
}

