% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareMETALS.range.multiAllele.R
\name{rareMETALS.formatData}
\alias{rareMETALS.formatData}
\title{Format data into analyzable format;}
\usage{
rareMETALS.formatData(raw.data)
}
\arguments{
\item{raw.data}{The data that is imported using seqminer;}
}
\value{
a list of formatted data;
}
\description{
Format data into analyzable format;
}

