% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rareMETALS.gene}
\alias{rareMETALS.gene}
\title{Meta-analysis of gene-level tests;}
\usage{
rareMETALS.gene(ANNO, score.stat.file, cov.file, gene, test = "GRANVIL",
  maf.cutoff, no.boot = 10000, alternative = c("two.sided", "greater",
  "less"), alpha = 0.05, ix.gold = 1, out.digits = 4,
  callrate.cutoff = 0, hwe.cutoff = 0, gene.file = "refFlat_hg19.txt.gz",
  max.VT = NULL)
}
\arguments{
\item{ANNO}{type of variants to be analyzed;}

\item{score.stat.file}{files of score statistics}

\item{cov.file}{covariance matrix files}

\item{test}{rare variant tests to be used}

\item{maf.cutoff}{MAF cutoff used to analyze variants}

\item{no.boot}{Number of bootstraps to be used. No need if asymptotics are used}

\item{alternative}{alternative hypothesis to be specified}

\item{alpha}{Significance threshold to determine the number of resampling. Set to 0 if analytic p-values are calculated.}

\item{ix.gold}{Gold standard population to align reference allele to}

\item{out.digits}{Number of digits used in the output}

\item{callrate.cutoff}{Cutoffs of call rate, lower than which will NOT be analyzed (labelled as missing)}

\item{hwe.cutoff}{Cutoffs of HWE p-values, lower than which will NOT be analyzed (labelled as missing)}

\item{max.VT}{The maximum number of thresholds used in VT; Setting max.VT to 10 can improve the speed for calculation without affecting the power too much. The default parameter is NULL, which does not set upper limit on the number of variable frequency threhsold.}
}
\value{
a list consisting of results
}
\description{
Meta-analysis of gene-level tests;
}

