% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareMETALS.parseAnno.R
\name{rareMETALS.parseAnno}
\alias{rareMETALS.parseAnno}
\title{A function to pass ANNOFULL from SEQMINER;}
\usage{
rareMETALS.parseAnno(x)
}
\arguments{
\item{x}{The input of the ANNOFULL from SEQMINER;}
}
\value{
A string with parsed information;
}
\description{
A function to pass ANNOFULL from SEQMINER;
}

