% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rareMETALS.range.core}
\alias{rareMETALS.range.core}
\title{#' Meta-analysis of gene-level tests by range;}
\usage{
rareMETALS.range.core(score.stat.file, cov.file, range, range.name,
  test = "GRANVIL", maf.cutoff = 1, alternative = c("two.sided",
  "greater", "less"), ix.gold = 1, out.digits = 4, callrate.cutoff = 0,
  hwe.cutoff = 0, max.VT = NULL)
}
\arguments{
\item{score.stat.file}{files of score statistics}

\item{cov.file}{covariance matrix files}

\item{range}{tabix range for each gene/region}

\item{range.name}{The name of the range,e.g. gene names can be used}

\item{test}{rare variant tests to be used}

\item{maf.cutoff}{MAF cutoff used to analyze variants}

\item{alternative}{alternative hypothesis to be specified}

\item{ix.gold}{Gold standard population to align reference allele to}

\item{out.digits}{Number of digits used in the output}

\item{callrate.cutoff}{Cutoffs of call rate, lower than which will NOT be analyzed (labelled as missing)}

\item{hwe.cutoff}{Cutoffs of HWE p-values}

\item{max.VT}{The maximum number of thresholds used in VT; Setting max.VT to 10 can improve the speed for calculation without affecting the power too much. The default parameter is NULL, which does not set upper limit on the number of variable frequency threhsold.}
}
\value{
a list consisting of results
}
\description{
#' Meta-analysis of gene-level tests by range;
}

