% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareMETALS.range.group.R
\name{rareMETALS.range.group}
\alias{rareMETALS.range.group}
\title{Meta-analysis of gene-level tests by range;}
\usage{
rareMETALS.range.group(score.stat.file, cov.file, range, range.name,
  test = "GRANVIL", refaltList, maf.cutoff = 1,
  alternative = c("two.sided", "greater", "less"), out.digits = 4,
  callrate.cutoff = 0, hwe.cutoff = 0, max.VT = NULL,
  correctFlip = TRUE, analyzeRefAltListOnly = TRUE, ...)
}
\arguments{
\item{score.stat.file}{files of score statistics}

\item{cov.file}{Covariance matrix files}

\item{range}{tabix range for each gene/region; Tabix range needs to be in the format of "1:123-1234". Quotation marks are necessary.}

\item{range.name}{The name of the range,e.g. gene names can be used;}

\item{test}{rare variant tests to be used}

\item{refaltList}{A list of reference, alternative allele, a vector of alternative allele frequencies; Specifically, the list should consist of}

\item{maf.cutoff}{MAF cutoff used to analyze variants; Default value is 1, i.e. no cutoffs are applied. MAFs are based upon the sample MAFs.}

\item{alternative}{alternative hypothesis to be specified; Only applicable to GRANVIL test;}

\item{out.digits}{Number of digits used in the output}

\item{callrate.cutoff}{Cutoffs of call rate, lower than which will NOT be analyzed (labelled as missing)}

\item{hwe.cutoff}{Cutoffs of HWE p-values}

\item{max.VT}{The maximum number of thresholds used in VT; For small p-values, the calculation of VT p-values can be very slow. Setting max.VT to 10 can improve the speed for calculation without affecting the power too much. The default parameter is NULL, which does not set upper limit on the number of variable frequency threhsold.}

\item{correctFlip}{Correct for flipped sites for score statistics and their covariance matrices}
}
\value{
a list consisting of results;
}
\description{
Meta-analysis of gene-level tests by range;
}

