% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareMETALS.range.multiAllele.R
\name{rareMETALS.range.multiAllele}
\alias{rareMETALS.range.multiAllele}
\title{Function to perform multi-allelic association test}
\usage{
rareMETALS.range.multiAllele(score.stat.file, cov.file, tabixRange, rangeName,
  maf.cutoff, rvTest)
}
\arguments{
\item{score.stat.file}{The file names for score statistics files;}

\item{cov.file}{The file names for covariance matrix files;}

\item{tabixRange}{The vectors of tabix ranges;}

\item{rangeName}{The name of the genes that are represented by the corresponding tabixRange;}

\item{maf.cutoff}{The MAF cutoffs;}

\item{rvTest}{The rare variant association tests to be performed. Options include GRANVIL, SKAT or VT;}
}
\value{
A matrix of results; each row being the result for a gene;
}
\description{
Function to perform multi-allelic association test
}

