% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareMETALS.single.group.R
\name{rareMETALS.single.group}
\alias{rareMETALS.single.group}
\title{Single variant meta-analysis}
\usage{
rareMETALS.single.group(score.stat.file, cov.file, range, refaltList,
  alternative = c("two.sided", "greater", "less"), callrate.cutoff = 0,
  hwe.cutoff = 0, correctFlip = TRUE, analyzeRefAltListOnly = TRUE)
}
\arguments{
\item{score.stat.file}{files of score statistics}

\item{cov.file}{covariance matrix files}

\item{range}{tabix range of variants to be analyzed}

\item{refaltList}{A list of reference alternative and positions of variants to be analyzed; Each variant in the dataset will be match against ref/alt alleles specified in refaltList; Only variants with matched ref and alt alleles can be included; we also need AF and af.diff.max to determine if the flips are due to strand issues or due to ref/alt alleles flips;}

\item{alternative}{alternative hypothesis to be specified}

\item{callrate.cutoff}{Cutoffs of call rate, lower than which will NOT be analyzed (labelled as missing)}

\item{hwe.cutoff}{Cutoffs of HWE p-values}

\item{correctFlip}{Correcting for flipped alleles; Default is TRUE; If FALSE, studies with incorrect REF/ALT alleles will be labelled as missing, and dropped from meta-analyses}

\item{analyzeRefAltListOnly}{Only analyze variants that are included in the refaltList; Default is TRUE; If FALSE, variant sites in the dataset but not specified in the refaltList will be labelled as missing and dropped from studies;}

\item{refaltList}{A list that contains pos, ref and alt for calibrating variant sites to have the right ref and alt alleles;}
}
\value{
a list consisting of results
}
\description{
Single variant meta-analysis
}

