% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rareMETALS.single}
\alias{rareMETALS.single}
\title{Single variant meta-analysis}
\usage{
rareMETALS.single(score.stat.file, cov.file, range,
  alternative = c("two.sided", "greater", "less"), ix.gold = 1,
  callrate.cutoff = 0, hwe.cutoff = 0)
}
\arguments{
\item{score.stat.file}{files of score statistics}

\item{cov.file}{covariance matrix files}

\item{range}{tabix range of variants to be analyzed}

\item{alternative}{alternative hypothesis to be specified}

\item{ix.gold}{Gold standard population to align reference allele to}

\item{callrate.cutoff}{Cutoffs of call rate, lower than which will NOT be analyzed (labelled as missing)}

\item{hwe.cutoff}{Cutoffs of HWE p-values}
}
\value{
a list consisting of results
}
\description{
Single variant meta-analysis
}

