% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{readData}
\alias{readData}
\title{read rvtests and raremetal format;}
\usage{
readData(score.stat.file)
}
\arguments{
\item{score.stat.file}{The score statistics file;}
}
\value{
the formatted stat;
}
\description{
read rvtests and raremetal format;
}
