% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvmeta.sim.R
\name{set.intersect}
\alias{set.intersect}
\title{Intersection of two sets;}
\usage{
set.intersect(set1, set2)
}
\arguments{
\item{set1}{The set 1}

\item{set2}{The set 2}
}
\value{
The intersection of set1 and set2;
}
\description{
Intersection of two sets;
}
