% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortPos.R
\name{sortPos}
\alias{sortPos}
\title{Sort genomic positions;}
\usage{
sortPos(pos)
}
\arguments{
\item{pos}{Genomic positions in the form of 1:12345;}
}
\value{
Return sorted genomic positions;
}
\description{
Sort genomic positions;
}

