% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareMETALS2.utils.R
\name{splitVecByChunkSize}
\alias{splitVecByChunkSize}
\title{This function implement how to split a vector into chunks of certain size}
\usage{
splitVecByChunkSize(vecIn, chunkSize)
}
\arguments{
\item{vecIn}{The input vector of any type;}

\item{chunkSize}{An integer for the chunk.size;}
}
\value{
A list with splitted chunks;
}
\description{
This function implement how to split a vector into chunks of certain size
}
