% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myUtils.R, R/rareMETALS2.utils.R
\name{toMinor}
\alias{toMinor}
\title{To minor: change number >.5 to 1-#}
\usage{
toMinor(vecIn)

toMinor(vecIn)
}
\value{
a vector with flipped number;

a vector with flipped number;
}
\description{
To minor: change number >.5 to 1-#

To minor: change number >.5 to 1-#
}
