% Generated by roxygen2 (4.0.0): do not edit by hand
\name{find.gene.chrpos}
\alias{find.gene.chrpos}
\title{Find the chromosomal position for genes;}
\usage{
find.gene.chrpos(candidate.gene)
}
\arguments{
\item{candidate.gene}{Name of candidate genes;}
}
\value{
a list consisiting of tabix, pos.start and pos.end;
}
\description{
Find the chromosomal position for genes;
}

