% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortPos.R
\name{genomic.dist}
\alias{genomic.dist}
\title{Get genomic distance between two variant positions;}
\usage{
genomic.dist(pos1, pos2)
}
\arguments{
\item{pos1}{Position for variant 1;}

\item{pos2}{Position for variant 2;}
}
\value{
Return variant positions;
}
\description{
Get genomic distance between two variant positions;
}

