% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.gene.R
\name{get.tabix.range}
\alias{get.tabix.range}
\title{change position to tabix range;}
\usage{
get.tabix.range(variant.vec)
}
\arguments{
\item{variant.vec}{The vector of variants of the format "1:12345";}
}
\value{
A vector of tabix ranges;
}
\description{
change position to tabix range;
}

